Remove AI-generated MCP slop
