+++ server/tool_policy.go
@@ -11,11 +11,10 @@ import (
 )
 
 // pluginServerOriginPrefix mirrors the wire format produced by
-// mcp.pluginServerOriginKey (mcp/user_clients.go:227) and the synthetic
-// origin keys assembled by filterToolsByConfig (mcp/client_manager.go:493)
-// and the Phase 3 external-endpoint synthetic config
-// (mcpserver/plugin_handlers.go:215). Kept as a package-private constant in
-// the server package because the mcp helper that constructs it is unexported.
+// mcp.pluginServerOriginKey and the synthetic origin keys assembled by
+// filterToolsByConfig and mcpserver plugin handlers. Kept as a package-private
+// constant in the server package because the mcp helper that constructs it is
+// unexported.
 const pluginServerOriginPrefix = "plugin://"
 
 // lookupToolPolicy is the pure decision function backing the
@@ -26,25 +25,20 @@ const pluginServerOriginPrefix = "plugin://"
 // Resolution order — first matching branch wins:
 //
 //  1. mcp.EmbeddedClientKey               → embedded server policy (with
-//                                            vetted-seed fallback when no
-//                                            persisted ToolConfigs).
+//     vetted-seed fallback when no
+//     persisted ToolConfigs).
 //  2. exact-match on cfg.Servers[i].BaseURL → remote MCP server policy.
 //  3. plugin://<pluginID> prefix          → cfg.PluginServers[i].PluginID
-//                                            lookup (M2 Phase 5 remediation;
-//                                            Phase 1 added PluginServers but
-//                                            did not wire the policy checker).
+//     lookup.
 //  4. fallthrough                         → ("ask", false). Means "do not
-//                                            auto-execute"; tool either was
-//                                            never surfaced (filter dropped
-//                                            it) or the origin is unknown.
+//     auto-execute"; tool either was
+//     never surfaced (filter dropped
+//     it) or the origin is unknown.
 //
 // Synthetic *mcp.ServerConfig for plugin entries hardcodes Enabled=true and
-// relies on the upstream `if !ps.Enabled { continue }` filter — mirrors
-// Phase 3's external-endpoint pattern at mcpserver/plugin_handlers.go:212-217
-// and filterToolsByConfig at mcp/client_manager.go:489-501. Propagating
-// ps.Enabled here would cause MCPServerConfig.GetToolPolicy to short-circuit
-// to ("ask", false) before consulting ToolConfigs, which would silently hide
-// every tool of a re-enabled plugin during a stale-cache window.
+// relies on the upstream `if !ps.Enabled { continue }` filter. Propagating
+// ps.Enabled here would cause MCPServerConfig.GetToolPolicy to short-circuit to
+// ("ask", false) before consulting ToolConfigs.
 func lookupToolPolicy(mcpCfg config.MCPConfig, serverBaseURL, toolName string) (string, bool) {
 	if serverBaseURL == mcp.EmbeddedClientKey {
 		toolConfigs := mcpCfg.EmbeddedServer.ToolConfigs
