+++ server/tool_policy_test.go
@@ -11,12 +11,8 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-// TestLookupToolPolicy_PluginServer_AutoRunEverywhere is the regression-pin
-// for M2 Phase 5 task #15. Before the fix, plugin tool calls fell through to
-// ("ask", false) because the policyChecker closure had no plugin:// branch,
-// forcing every plugin tool to require manual approval regardless of admin
-// policy. This test asserts that an admin-set auto_run_everywhere policy on
-// a plugin tool is honored.
+// TestLookupToolPolicy_PluginServer_AutoRunEverywhere verifies that an
+// admin-set auto_run_everywhere policy on a plugin tool is honored.
 func TestLookupToolPolicy_PluginServer_AutoRunEverywhere(t *testing.T) {
 	const pluginID = "com.mattermost.plugin-demo"
 	const toolName = "com_mattermost_plugin-demo__add_two_numbers"
@@ -38,8 +34,7 @@ func TestLookupToolPolicy_PluginServer_AutoRunEverywhere(t *testing.T) {
 
 	assert.Equal(t, config.MCPToolPolicyAutoRunEverywhere, policy,
 		"plugin tool with admin-set auto_run_everywhere must report that policy "+
-			"(regression: M2 Phase 1 added PluginServerConfig.ToolConfigs but did "+
-			"not extend the policyChecker closure)")
+			"through the plugin:// policy branch")
 	assert.True(t, enabled, "plugin tool with admin Enabled=true must report enabled=true")
 }
 
@@ -111,7 +106,6 @@ func TestLookupToolPolicy(t *testing.T) {
 				PluginID: pluginID,
 				Name:     "Demo Plugin",
 				Enabled:  true,
-				// ToolConfigs intentionally empty — pre-M2 default-allow.
 			}},
 		}
 		policy, enabled := lookupToolPolicy(cfg, pluginOrigin, pluginToolName)
@@ -148,10 +142,6 @@ func TestLookupToolPolicy(t *testing.T) {
 		assert.Equal(t, config.MCPToolPolicyAsk, policy)
 		assert.False(t, enabled)
 	})
-
-	// --- Regression pins for the embedded and remote branches. These exist so
-	// future edits to lookupToolPolicy cannot silently regress the paths that
-	// already worked before the M2 Phase 5 fix.
 
 	t.Run("remote server auto_run_everywhere -> propagates", func(t *testing.T) {
 		cfg := config.MCPConfig{
