+++ webapp/src/components/system_console/mcp_tools_viewer.test.tsx
@@ -31,9 +31,7 @@ jest.mock('react-intl', () => {
     };
 });
 
-// Mock the client module — Phase 4 tests assert on PUT bodies via the
-// updatePluginServer mock. Other client functions are stubbed enough to
-// make the component render.
+// Mock the client module; tests assert on PUT bodies via updatePluginServer.
 jest.mock('../../client', () => ({
     __esModule: true,
     getMCPTools: jest.fn(),
@@ -122,8 +120,6 @@ describe('MCPToolsViewer — plugin branch', () => {
         // Expand the server row so the tool list renders.
         fireEvent.click(screen.getByText('Demo Plugin'));
 
-        // The policy dropdowns are rendered for the plugin row's tools.
-        // None should be disabled — Phase 4 re-enabled the dropdown for plugin rows.
         const selects = screen.getAllByRole('combobox');
         expect(selects.length).toBeGreaterThanOrEqual(1);
         for (const sel of selects) {
